+++ packages/@tenoxui-core/vite.config.ts
@@ -5,4 +5,15 @@ export default createConfig({
   entry: 'src/index.ts',
   formats: ['es', 'iife', 'cjs', 'umd'],
   sourcemap: false
+  /*
+  rollupOptions: {
+    output: {
+      exports: 'named',
+      globals: {
+        '@tenoxui/moxie': '__tenoxui_moxie__'
+      }
+    },
+    external: ['@tenoxui/moxie']
+  }
+  */
 })